requirements.txt
# aiogram 3.x (пока в бета-версии, например 3.0.0b7)
aiogram==3.0.0b7

# Подключение к MongoDB
pymongo==4.3.3

# Для загрузки переменных окружения из .env
python-dotenv==0.21.1

# Для HTTP-запросов (Market.CSGO, Waxpeer и т.п.)
requests==2.28.2

# Для кэширования или очередей, если понадобится
redis==4.5.1

# Для работы с WebSocket (реальное время), если понадобится
websocket-client==1.5.1
